+++ packages/eslint-config-tester/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.7](https://github.com/udemy/js-tooling/compare/eslint-config-tester@3.0.6...eslint-config-tester@3.0.7) (2019-10-16)
+
+**Note:** Version bump only for package eslint-config-tester
+
+
+
+
+
 ## [3.0.6](https://github.com/udemy/js-tooling/compare/eslint-config-tester@3.0.5...eslint-config-tester@3.0.6) (2019-09-26)
 
 **Note:** Version bump only for package eslint-config-tester