+++ packages/prettier-config-udemy-website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/udemy/js-tooling/compare/prettier-config-udemy-website@1.0.6...prettier-config-udemy-website@1.0.7) (2019-10-16)
+
+**Note:** Version bump only for package prettier-config-udemy-website
+
+
+
+
+
 ## [1.0.6](https://github.com/udemy/js-tooling/compare/prettier-config-udemy-website@1.0.5...prettier-config-udemy-website@1.0.6) (2019-09-26)
 
 **Note:** Version bump only for package prettier-config-udemy-website